Move truncate words to text utils
